# lib/api-zod/src/generated/types/authSession.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * XpressPro FX Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { AuthSessionMerchantStatus } from "./authSessionMerchantStatus";
import type { AuthSessionRole } from "./authSessionRole";
import type { User } from "./user";

export interface AuthSession {
  user: User | null;
  role: AuthSessionRole;
  isDemo: boolean;
  /** True when the user dismissed the mandatory connect-wallet interstitial. */
  walletSkipped: boolean;
  /** True when the user is an approved P2P merchant. */
  isMerchant: boolean;
  /** Status of the user's most recent P2P merchant application, or null when none has been submitted. */
  merchantStatus: AuthSessionMerchantStatus;
}