# openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: XpressPro FX Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /coinbase/webhook:
    post:
      operationId: coinbaseWebhook
      tags: [coinbase]
      summary: Coinbase transaction webhook (purchase completion events)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CoinbaseWebhookEvent"
      responses:
        "200":
          description: Webhook processed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
...
  /auth/signup:
    post:
      operationId: signup
      tags: [auth]
      summary: Register a new user (sends OTP, does not create session)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignupRequest"
      responses:
        "200":
          description: OTP challenge issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthOtpChallenge"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Log in with email and password (admins authenticate immediately, others receive an OTP)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Either an authenticated session (admins) or an OTP challenge (regular users)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResult"
...
  /auth/verify-otp:
    post:
      operationId: verifyOtp
      tags: [auth]
      summary: Verify a 6-digit OTP and complete the auth flow
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyOtpRequest"
      responses:
        "200":
          description: Authenticated session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSession"
  /auth/resend-otp:
    post:
      operationId: resendOtp
      tags: [auth]
      summary: Generate and resend a fresh OTP for a pending challenge
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResendOtpRequest"
      responses:
        "200":
          description: New OTP issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthOtpChallenge"
...
  /moonpay/webhook:
    post:
      operationId: moonpayWebhook
      tags: [moonpay]
      summary: MoonPay transaction webhook (purchase completion events)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MoonpayWebhookEvent"
      responses:
        "200":
          description: Webhook processed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
...
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        email:
          type: string
        fullName:
          type: string
        country:
          type: string
        kycVerified:
          type: boolean
        avatarUrl:
          type: string
        createdAt:
          type: string
        selectedManagerId:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        merchant:
          type: boolean
          description: True when the user is an approved P2P merchant.
        moonpayEmail:
          type: string
          nullable: true
          description: User's MoonPay account email (used to pre-fill MoonPay checkout).
        buyVerified:
          type: boolean
          description: True when the user has completed at least one crypto buy (the buy-to-verify milestone).
      required: [id, username, email, fullName, country, kycVerified, createdAt, buyVerified]
...
    AuthSession:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
          nullable: true
        role:
          type: string
          enum: [user, admin, demo]
        isDemo:
          type: boolean
        walletSkipped:
          type: boolean
          description: True when the user dismissed the mandatory connect-wallet interstitial.
        isMerchant:
          type: boolean
          description: True when the user is an approved P2P merchant.
        merchantStatus:
          type: string
          enum: [pending, approved, rejected]
          nullable: true
          description: Status of the user's most recent P2P merchant application, or null when none has been submitted.
      required: [user, role, isDemo, walletSkipped, isMerchant, merchantStatus]
    SignupRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
        fullName:
          type: string
        country:
          type: string
        referralCode:
          type: string
          nullable: true
      required: [email, password, fullName, country]
    LoginRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required: [email, password]
    AuthOtpChallenge:
      type: object
      description: Returned by signup/login/resend-otp; client must call /auth/verify-otp with the code.
      properties:
        status:
          type: string
          enum: [otp_required]
        email:
          type: string
        intent:
          type: string
          enum: [signup, login]
        expiresInSeconds:
          type: integer
        message:
          type: string
      required: [status, email, intent, expiresInSeconds, message]
    AuthSessionAuthenticated:
      type: object
      description: An authenticated session returned directly from /auth/login when the user is an admin.
      allOf:
        - $ref: "#/components/schemas/AuthSession"
        - type: object
          properties:
            status:
              type: string
              enum: [authenticated]
          required: [status]
    LoginResult:
      oneOf:
        - $ref: "#/components/schemas/AuthSessionAuthenticated"
        - $ref: "#/components/schemas/AuthOtpChallenge"
      discriminator:
        propertyName: status
        mapping:
          authenticated: "#/components/schemas/AuthSessionAuthenticated"
          otp_required: "#/components/schemas/AuthOtpChallenge"
...
    VerifyOtpRequest:
      type: object
      properties:
        email:
          type: string
        code:
          type: string
          minLength: 6
          maxLength: 6
      required: [email, code]
    ResendOtpRequest:
      type: object
      properties:
        email:
          type: string
      required: [email]
...
    OkResponse:
      type: object
      properties:
        ok:
          type: boolean
      required: [ok]
...
    MoonpayWebhookEvent:
      type: object
      description: A subset of the MoonPay webhook payload we react to.
      properties:
        type:
          type: string
          description: Event type, e.g. "transaction_updated" or "transaction_completed".
        data:
          type: object
          description: The MoonPay transaction object.
          properties:
            id:
              type: string
            status:
              type: string
              description: e.g. pending, completed, failed.
            walletAddress:
              type: string
              nullable: true
            currency:
              type: object
              nullable: true
              properties:
                code:
                  type: string
            cryptoTransactionId:
              type: string
              nullable: true
            quoteCurrencyAmount:
              type: number
              nullable: true
              description: Amount of crypto purchased.
            baseCurrencyAmount:
              type: number
              nullable: true
              description: Fiat amount spent.
            baseCurrencyCode:
              type: string
              nullable: true
            externalCustomerId:
              type: string
              nullable: true
              description: Platform user id we passed in the checkout URL.
            externalTransactionId:
              type: string
              nullable: true
              description: Server-generated id we passed at /moonpay/initiate; used to look up the original destination so the webhook only credits platform wallets when the user picked one.
          required: [id, status]
      required: [type, data]
...
    CoinbaseWebhookEvent:
      type: object
      properties:
        type:
          type: string
        data:
          type: object
          properties:
            id:
              type: string
            status:
              type: string
            walletAddress:
              type: string
              nullable: true
            currency:
              type: object
              nullable: true
              properties:
                code:
                  type: string
            quoteCurrencyAmount:
              type: number
              nullable: true
            baseCurrencyAmount:
              type: number
              nullable: true
            baseCurrencyCode:
              type: string
              nullable: true
            externalCustomerId:
              type: string
              nullable: true
            externalTransactionId:
              type: string
              nullable: true
          required: [id, status]
      required: [type, data]
...